sprint12

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -8,17 +8,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Genre {
-    private int id;
+    private Long id;
     private String name;
 
-    public static Genre fromId(int id) {
-        return switch (id) {
-            case 1 -> new Genre(1, "COMEDY");
-            case 2 -> new Genre(2, "DRAMA");
-            case 3 -> new Genre(3, "CARTOON");
-            case 4 -> new Genre(4, "THRILLER");
-            case 5 -> new Genre(5, "DOCUMENTARY");
-            case 6 -> new Genre(6, "ACTION");
+    public static Genre fromId(long id) {
+        return switch ((int) id) {
+            case 1 -> new Genre(1L, "COMEDY");
+            case 2 -> new Genre(2L, "DRAMA");
+            case 3 -> new Genre(3L, "CARTOON");
+            case 4 -> new Genre(4L, "THRILLER");
+            case 5 -> new Genre(5L, "DOCUMENTARY");
+            case 6 -> new Genre(6L, "ACTION");
             default -> throw new IllegalArgumentException("Invalid genre id: " + id);
         };
     }
